[pace] Added piecewise polynomial support to architecture
- Added support for piecewise polynomial approximation via PACE mode.
- PACE can be enabled by setting `pace_mode`; otherwise, it remains disabled.
- The PACE logic is conditionally included only when `PACE_ENABLED` is defined.
- X-buffers are used to supply breakpoints and coefficients for approximation.
- Input and output data are streamed using dedicated FIFOs to perform element-wise processing.
- Verified implementation using 8x8 configuration.
- In each row of 8 CEs:
  - 3 CEs are used in the BST to generate the partition ID.
  - 4 CEs are used to implement a polynomial of degree 4.
  - CE-3 is bypassed to support the required 5 coefficients, with CE-3 forwarding the coefficient to the next CE.
- `pace.py` (golden model) generates stimuli to verify the PACE extension.
- It also generates a debug file to trace datapath-level execution for easier debugging.

Usage:
  make golden sw-all hw-clean hw-all target=vsim gui=1 OP=pace

+++ golden-model/FP16/scripts/gemm.py
@@ -38,10 +38,10 @@ f = open(args.file_name, "w")
 
 # We want to perform a GEMM, of the kind Z = Y + X*W
 # Test Matrices
-X = torch.rand(m_size, n_size).half()
-W = torch.rand(n_size, k_size).half()
-Y = torch.rand(m_size, k_size).half()
-Z = torch.rand(m_size, k_size).half()
+X = torch.rand(m_size, n_size)
+W = torch.rand(n_size, k_size)
+Y = torch.rand(m_size, k_size)
+Z = torch.rand(m_size, k_size)
 
 print("\nInput Data: ")
 print("\nX is: ", X, X.shape, X.dtype)
